+++ scripts/conab/fetch_reference_samples.sh
@@ -15,6 +15,8 @@ fetch() {
 
 fetch "LevantamentoGraos.txt" "https://portaldeinformacoes.conab.gov.br/downloads/arquivos/LevantamentoGraos.txt"
 fetch "SerieHistoricaGraos.txt" "https://portaldeinformacoes.conab.gov.br/downloads/arquivos/SerieHistoricaGraos.txt"
+fetch "OfertaDemanda.txt" "https://portaldeinformacoes.conab.gov.br/downloads/arquivos/OfertaDemanda.txt"
+fetch "PrecosSemanalUF.txt" "https://portaldeinformacoes.conab.gov.br/downloads/arquivos/PrecosSemanalUF.txt"
 fetch "LevantamentoCana.txt" "https://portaldeinformacoes.conab.gov.br/downloads/arquivos/LevantamentoCana.txt"
 fetch "SerieHistoricaCana.txt" "https://portaldeinformacoes.conab.gov.br/downloads/arquivos/SerieHistoricaCana.txt"